Improve Database DI and code structure.
