+++ be/src/main/scala/io/nariok/services/BlogService.scala
@@ -2,19 +2,22 @@ package io.nariok.services
 
 import com.google.inject.Inject
 import io.nariok.models.blog.Article
-import io.nariok.repositories.MysqlRepository
+import io.nariok.repositories.DatabaseRepository
+import io.nariok.repositories.mappers.{ArticleMapper, BlogContentMapper}
 import io.nariok.routes.JsonSupport
 
 import scala.concurrent.ExecutionContext
 
-object Database {
-  val database = new MysqlRepository
-  database.connect()
+trait BlogService {
+  def getArticles(): List[Article]
+}
+
+class Database @Inject()(database: DatabaseRepository) {
   val articlesSql = "SELECT id, title, created, modified FROM blog " +
     "ORDER BY created DESC, modified DESC LIMIT 5"
-  val blogs      = database.executeQuery(articlesSql, None, MysqlRepository.toArticle)
+  val blogs      = database.executeQuery(articlesSql, None, ArticleMapper.toArticle)
   val contentSql = "SELECT id, content, type, blog_id, alttext, align FROM blog_content"
-  val contents   = database.executeQuery(contentSql, None, MysqlRepository.toBlogContent)
+  val contents   = database.executeQuery(contentSql, None, BlogContentMapper.toBlogContent)
 
   def articles(): List[Article] =
     blogs.map(blog => {
@@ -26,15 +29,11 @@ object Database {
         })
       article
     })
-  database.disconnect()
 }
 
-class BlogServiceImpl @Inject()()(implicit private val executionContext: ExecutionContext)
+class BlogServiceImpl @Inject()(private val databaseRepository: DatabaseRepository)(
+    implicit private val executionContext: ExecutionContext)
     extends BlogService
     with JsonSupport {
-  override def getArticles(): List[Article] = Database.articles()
-}
-
-trait BlogService {
-  def getArticles(): List[Article]
+  override def getArticles(): List[Article] = new Database(databaseRepository).articles()
 }
